fix warning and add resend time by rto.

+++ common/Rto.cpp
@@ -19,11 +19,11 @@ CRtoImpl::~CRtoImpl() {
 void CRtoImpl::SetRttTime(uint64_t rtt) {
     std::unique_lock<std::mutex> lock(_mutex);
     if (_srtt == 0) {
-        _srtt = rtt;
-        _mdev = rtt >> 1;
+        _srtt = (uint32_t)rtt;
+        _mdev = (uint32_t)rtt >> 1;
 
     } else {
-        int32_t err = rtt - _srtt;
+        int32_t err = (uint32_t)(rtt - _srtt);
         _srtt = _srtt + G(err);
         _mdev = _mdev + H((int32_t)std::abs(err) - _mdev);
     }
